pyproject.toml
[project]
name = "rizzk-terminal"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "dash>=2,<3",
  "plotly>=5,<6",
  "flask>=3,<4",
  "flask-caching>=2,<3",
  "pandas",
  "numpy",
  "pandas-ta",
  "feedparser",
  "python-dotenv>=1.0",
  "waitress",
  "sqlalchemy>=2,<3",
  "psycopg2-binary",
  "diskcache",
  "sentry-sdk",
  "ruff",
  "pre-commit",
  "pytest",
  "requests"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["rizzk*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.ruff.lint.isort]
known-first-party = ["rizzk"]
